backend/pyproject.toml
[project]
name = "saraswati-backend"
version = "0.1.0"
description = "Saraswati AI — FastAPI Backend"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.5.0",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "neo4j>=5.20.0",
    "google-genai>=0.8.0",
    "arxiv>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
